+++ src/events.js
@@ -1,5 +1,5 @@
-import { DivIcon, Marker, Point, Polygon, Polyline, DomEvent } from 'leaflet';
-import { cls, isTrustedEvent } from './utils';
+import { DivIcon, Marker, Point, Polygon, Polyline, DomEvent, LayerGroup } from 'leaflet';
+import { cls, isTrustedEvent, CLICK_EVT } from './utils';
 import { addEndClickArea, removeEndClickArea } from './drawing-pane';
 
 function doNothingHandler(event) {
@@ -13,18 +13,24 @@ function doNothingHandler(event) {
  */
 export function onAddPoint(event) {
   // Default behavior while dragging
-  if (this.map_moving) {
+  if (this.mapMoving) {
     return;
   }
-  const { index = null } = event;
+  const map = this.getMap();
   let { clientX, clientY } = event;
   // Touch device
   if (clientX === undefined && clientY === undefined) {
-    const touch = event.touches[0];
+    const touch = event.changedTouches[0];
     clientX = touch.clientX;
     clientY = touch.clientY;
   }
-  const map = this.getMap();
+  // We were drawing a rect, so we don't perform the canonical action ans we stop here
+  if (this.rectDrawing) {
+    map.fire('as:dragging-rect-end');
+    return;
+  }
+
+  const { index = null } = event;
   const container = map.getContainer();
   const bbox = container.getBoundingClientRect();
   const x = clientX - bbox.left;
@@ -68,7 +74,7 @@ export function onAddPoint(event) {
       });
     })(this.markers.length)
   );
-  marker.on('click', doNothingHandler);
+  marker.on(CLICK_EVT, doNothingHandler);
   marker.addTo(map);
   map.fire('as:marker-add', newEdge);
   // If this point as not been added at the end, we need to update even handlers HOC params to update index
@@ -76,8 +82,8 @@ export function onAddPoint(event) {
     for (let i = index + 1; i < this.markers.length; i++) {
       this.markers[i].marker.off('drag');
       this.markers[i].marker.on('drag', _onMarkerDrag(i));
-      this.markers[i].marker.off('click');
-      this.markers[i].marker.on('click', doNothingHandler);
+      this.markers[i].marker.off(CLICK_EVT);
+      this.markers[i].marker.on(CLICK_EVT, doNothingHandler);
       this.markers[i].marker.off('dblclick');
       this.markers[i].marker.on('dblclick', (event) => {
         map.fire('as:marker-remove', {
@@ -159,8 +165,8 @@ export function onRemoveMarker({ index = 0 }) {
   for (let i = index; i < this.markers.length; i++) {
     this.markers[i].marker.off('drag');
     this.markers[i].marker.on('drag', onMarkerDrag.bind(this)(i));
-    this.markers[i].marker.off('click');
-    this.markers[i].marker.on('click', doNothingHandler);
+    this.markers[i].marker.off(CLICK_EVT);
+    this.markers[i].marker.on(CLICK_EVT, doNothingHandler);
     this.markers[i].marker.off('dblclick');
     this.markers[i].marker.on('dblclick', (event) => {
       event.originalEvent.stopPropagation();
@@ -192,7 +198,7 @@ export function onUpdatePolygon() {
       className: 'drawing-area-poligon',
     }
   );
-  polygon.on('click', (ev) => {
+  polygon.on(CLICK_EVT, (ev) => {
     DomEvent.stopPropagation(ev);
   });
 
@@ -236,10 +242,10 @@ export function onUpdateGhostPoints() {
         point,
         marker,
       };
-      marker.on('click', doNothingHandler);
+      marker.on(CLICK_EVT, doNothingHandler);
       marker.on('dblclick', doNothingHandler);
-      marker.on('drag', onGhostMarkerDrag.bind(this)(ghostMarkers.length));
       marker.on('dragstart', onGhostMarkerDragStart.bind(this)());
+      marker.on('drag', onGhostMarkerDrag.bind(this)(ghostMarkers.length));
       marker.on('dragend', onGhostMarkerDragEnd.bind(this)(ghostMarkers.length));
       ghostMarkers.push(newGhostMarker);
       marker.addTo(map);
@@ -249,6 +255,8 @@ export function onUpdateGhostPoints() {
 
 export function onPolygonCreationEnd() {
   const map = this.getMap();
+  // We'll forcly enable the drag on the map (which will be just of the draw pane)
+  map.dragging.enable();
   map.removeLayer(this.closeLine);
   this.closeLine = null;
   // Remove style for the final marker icon
@@ -259,19 +267,30 @@ export function onPolygonCreationEnd() {
   removeEndClickArea(this);
 }
 
+/**
+ * Plugin activation (button pressed)
+ */
 export function onActivate(event) {
   // Non-sense, but required on Safari. Probably related to https://github.com/Leaflet/Leaflet/issues/7255
   if (!isTrustedEvent(event)) {
     return;
   }
+  const map = this.getMap();
+  // storing if the dragging is enabled in the map
+  this._dragStatus = map.dragging._enabled;
+
   // Let leave to custom callback any call to preventDefault, which will block normal button behaviors
   // event.preventDefault();
   event.stopPropagation();
-  const map = this.getMap();
   event.target.blur();
   // if current state is active, we need to deactivate
   const activeState = this.options.active || this.phase === 'adjust';
   if (activeState) {
+    // Disable dragging if it was enabled before the activation
+    if (!this._dragStatus) {
+      map.dragging.disable();
+    }
+
     // Calling user's specific event handler
     this.options.onButtonDeactivate(this.polygon, this, event);
     if (!event.defaultPrevented) {
@@ -281,6 +300,8 @@ export function onActivate(event) {
     // Calling user's specific event handler
     this.options.onButtonActivate(this, event);
     if (!event.defaultPrevented) {
+      // When activating the plugin we'll disable dragging temporarely (in case we are drawing a rect)
+      map.dragging.disable();
       this.activateButton.classList.add('active');
       map.getContainer().classList.add('drawing-area');
       this.setPhase('draw', true);
@@ -359,4 +380,56 @@ export function onGhostMarkerDragEnd(index) {
     };
     map.fire('as:point-add', fakeEvent);
   };
+}
+
+export function onMouseMove(event) {
+  if (!this.mapMoving && this.markers.length === 0 && event.which !== 0 && event.buttons === 1) {
+    const map = this.getMap();
+    if (!this.rectDrawing) {
+      this.rectDrawStart = [event.clientX, event.clientY];
+      const props = {
+        weight: 2,
+        color: '#8B4513',
+        className: 'rect-progress-line',
+        opacity: 0.6,
+      };
+      this.draggingRect = new LayerGroup()
+        .addLayer(new Polyline([], props))
+        .addLayer(new Polyline([], props))
+        .addTo(map);
+      this.rectDrawing = true;
+      return;
+    }
+    this.rectDrawEnd = [event.clientX, event.clientY];
+    const pointA = [this.rectDrawStart[0], event.clientY];
+    const pointB = [event.clientX, this.rectDrawStart[1]];
+
+    const layers = this.draggingRect.getLayers();
+    layers[0].setLatLngs([
+      map.mouseEventToLatLng({ clientX: this.rectDrawEnd[0], clientY: this.rectDrawEnd[1] }),
+      map.mouseEventToLatLng({ clientX: pointA[0], clientY: pointA[1] }),
+      map.mouseEventToLatLng({ clientX: this.rectDrawStart[0], clientY: this.rectDrawStart[1] }),
+    ]);
+    layers[1].setLatLngs([
+      map.mouseEventToLatLng({ clientX: this.rectDrawEnd[0], clientY: this.rectDrawEnd[1] }),
+      map.mouseEventToLatLng({ clientX: pointB[0], clientY: pointB[1] }),
+      map.mouseEventToLatLng({ clientX: this.rectDrawStart[0], clientY: this.rectDrawStart[1] }),
+    ]);
+  }
+}
+
+export function onDraggingRectEnd() {
+  this.rectDrawing = false;
+  const vertex1 = { clientX: this.rectDrawStart[0], clientY: this.rectDrawStart[1] };
+  const vertex2 = { clientX: this.rectDrawEnd[0], clientY: this.rectDrawStart[1] };
+  const vertex3 = { clientX: this.rectDrawEnd[0], clientY: this.rectDrawEnd[1] };
+  const vertex4 = { clientX: this.rectDrawStart[0], clientY: this.rectDrawEnd[1] };
+  onAddPoint.bind(this)(vertex1);
+  onAddPoint.bind(this)(vertex2);
+  onAddPoint.bind(this)(vertex3);
+  onAddPoint.bind(this)(vertex4);
+  this._map.fire('as:creation-end');
+  this.rectDrawStart = null;
+  this.rectDrawEnd = null;
+  this.draggingRect.removeFrom(this._map);
 }